Trabajo en progreso, comunicacion master/slave. Se mejora interfaz gracica, se agrega soporte para BT.
- Se agrega libreira AltSoftSerial (dependencia de BLE).
- Se agrega libreria RedBearLab BLEmini.
- Se completa interfaz grafica de aplicación Qt (quedan pendientes varias
  funcionalidades).
- Se agrega release 0.9 de aplicacion Qt para windows.

+++ Firmware/libraries/Utils/Utils.h
@@ -29,7 +29,7 @@ public:
      * Put at the given address in the given memory pointer, the bytes representation of a given
      * value (by the data type of the value).
      * */
-    static void putInMemory(byte *mem, byte address, uint16_t value);
+    static void putInMemory(byte *mem, uint16_t address, uint16_t value);
 };
 
 #endif // __H_UTILS_
